Changes when ROMFS files are built
