+++ utils/Cargo.toml
@@ -25,10 +25,5 @@ serde_json = "1"
 [patch.'https://github.com/jam1garner/smash-arc']
 smash-arc = "0.5"
 
-[build-dependencies]
-hdr-macros = { path = "../hdr-macros" }
-prc-rs = { git = "https://github.com/ultimate-research/prc-rs", features = ["indexmap-std", "xml-feat"] }
-walkdir = "2.3.2"
-
 [features]
 no-offset-search = []
